+++ tests/testthat/test-round-trip.R
@@ -22,23 +22,23 @@ test_that("roundtrip works", {
 test_that("roundtrip to file works", {
   file <- tempfile()
   dat <- mtcars
-  zstd_serialize(dat, file = file)
+  zstd_serialize(dat, dst = file)
   expect_identical(dat, zstd_unserialize(file))
   
   dat <- iris
-  zstd_serialize(dat, file = file)
+  zstd_serialize(dat, dst = file)
   expect_identical(dat, zstd_unserialize(file))
   
   dat <- sample(1e6)
-  zstd_serialize(dat, file = file)
+  zstd_serialize(dat, dst = file)
   expect_identical(dat, zstd_unserialize(file))
   
   dat <- function(x) {3 + 7 + x}
-  zstd_serialize(dat, file = file)
+  zstd_serialize(dat, dst = file)
   expect_equal(dat, zstd_unserialize(file))
   
   # Multithreading
   dat <- sample(1e6)
-  zstd_serialize(dat, file = file, num_threads = 2)
+  zstd_serialize(dat, dst = file, num_threads = 2)
   expect_identical(dat, zstd_unserialize(file))
 })